Refactor into separate mojos and an abstract one to support easy invocation using maven sitegen:preprocess-adoc and sitegen:naturalize-adoc
